Renamed packages, edited .classpath and bnd.bnd, updated references to old package in OidcClientConfigImpl, and renamed packageinfo to package-info.java
