--- dev/io.openliberty.security.oidcclientcore.internal/src/io/openliberty/security/oidcclientcore/internal/discovery/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
